+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # ChangeLog
 
-* fix geckolib model offset
-
+* fix fluid filter require specific amount